+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.9] - 2026-04-21
 ### Added
 - Video download button
 - Subtitle download button
@@ -87,7 +88,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Load subtitle from URL
 
 
-[Unreleased]: https://github.com/sepandhaghighi/neovid/compare/v0.8...main
+[Unreleased]: https://github.com/sepandhaghighi/neovid/compare/v0.9...main
+[0.9]: https://github.com/sepandhaghighi/neovid/compare/v0.8...v0.9
 [0.8]: https://github.com/sepandhaghighi/neovid/compare/v0.7...v0.8
 [0.7]: https://github.com/sepandhaghighi/neovid/compare/v0.6...v0.7
 [0.6]: https://github.com/sepandhaghighi/neovid/compare/v0.5...v0.6